Don't hide morphology for languages with no INFIX_MARK
Currently "w/something", which is tokenized as "w/ something", is
combined back when display_morphology==0. Since the links to "w/"
cannot be removed, a null word ("\0") is left in its original
position, which may not be expected by the user.
Also, contracted words are never combined back.
So it may be a good idea not combine back also "w/".

Implement that by setting HIDE_MORPHO to false for languages that don't
use INFIX_MARK.