tests.py: Fix broken library locale setup testing
Fix "test_dictionary_locale_definition":

The ability of the library to use the dictionary-defined locale is
tested here by setting the user domain to tr_TR.UTF-8 and checking that
there is still a full parse (which is expected only if "I" gets
lowercased to "i", instead of to "ı").

However, there is a big in the test code:
locale.setlocale(locale.LC_CTYPE, tr_locale) returns the NEW locale,
so the locale is actually not restored after the test and remains
tr_TR.UTF-8.

This leads to a very strange bug: The C library "toupper" table is
strangely and incorrectly patched by python2 - the uppercase code of "i"
is set to itself! Here is a debug printout:
Current locale: en_US.UTF-8
toupper hij: HIJ
Turkish locale: tr_TR.UTF-8
toupper hij: HiJ
Restored locale: tr_TR.UTF-8
toupper hij: HiJ

As a result, when I added a test using the "he" dict (that uses he_IL.UTF-8)
the follwing problem happens in the library due to the then bogus
"toupper" table:
..Debug: Dictionary "he/4.0.dict": Locale "he_iL.UTF-8" unknown
(note the lowercase "i").

After solving the said incorrect use of locale.setlocale(),
it was found that on python2 it returns the system locale instead of
the user one, so it is restored incorrectly. On python3 all is fine
(on Linux).

Summary:
- Fix finding the original locale.
- Disable the test on python2.
- Add (commented out) debug code.