MSVC: Move link-features.h to msvc/link-grammar/
The current MSVC build generates it in link-grammar/ directly under the
source directory. If later a *nix compilation is done, it is wrongly
picked up.

Fix it by generating it in msvc/link-grammar. Note that using
a subdirectory link-grammar/ under msvc/ is needed because it is
included as "link-grammar/link-features.h".

See discussion at PR #788.