Windows: Convert argv from the startup locale to UTF-8
When compiled by MSVC or MINGW64, non-ASCII character in link-parser's
argument such as "-!café.n" remain in the startup locale encoding but
the program expects utf-8. This causes to a crash in the case of MSVC
and to using a garbage value in the case of MINGW64.

(This problem doesn't exist on Cygwin even when running under Windows.)

Solve it by converting argv to utf-8.