default_error_handler(): Use fputs() instead of fprintf()
For some reason, on MINGW64 fprintf() doesn't follow fd redirection in
Python.  The work around here is to use fputs(), which doesn't have this
problem.  (It is also slightly more efficient because there is no format
to interpret).

Without this fix, the tests.py tests that use divert_start(), fail.