MSVC: Add support for DICTIONARY_DIR
Support absolute DICTIONARY_DIR path and also a relative
path (relative to the location of the program).

1. If it is relative, concatenate it to the program location in order
to find the system dictionary directory.
2. If it is absolute and contain a drive, or is a UNC path, it is
the system dictionary directory.
3. If it is absolute and doesn't contain a drive (and is not a UNC
path), concatenate it to the program location drive (or UNC host
specification).

Also, leave the custom dictionary directory undefined (previously it
was used to specify the system dictionary location, which is not a good
idea since it is the first path which is used and not the last resort
one).