Add regex package selection + pcre2 support
1. By default, use the first POSIX regex implementation found (both
header file and library):

libpcre2-posix
libtre
libc

In order to ensure that regcomp() etc. will not get accidentally bound
to libc (as may happen in the language bindings, e.g. when used from
python), the source code of pcre2posix.c (from the pcre2 package) is
used (BSD license). This will hopefully not be needed starting from the
next pcre2 version (10.33) and is arranged to automatically get ifdef'ed
out then.

2. A particular regex library from this list can be configured by
  --with-regexlib=tre|pcre2|c.

3. The original implementation of this patch included also selecting
the use of c++ regex (to natively support Windows). However, it got
removed from the implementation because it has a problem in fully
supporting Unicode.

4. The check for libtre is now done using PKG_CHECK_MODULES.

5. Define HAVE_REGEX_H on WIN32 to pick up regex.h.