configure: Unify CFLAGS and CXXFLAGS for various systems
It turns out the values currently used for Linux/*BSD cam also
be sued for the other supported OSs.

Use #" at the end of the added assignment lines that have values with
an equal sign, as a counter measure to vim syntax highlighting confusion
that otherwise occurs when it encounters odd number of such lines.