manpage: Specify system data directory with no replacement
Currently it is mentioned using the string %DATA_DIR% that get replaced
by "sed" in the manpage Makefile to the configured DICTIONARY_DIR. This
is unfortunate because it causes a confusion in case a manpage is used
directly from the source directory (happens in many online manpage
repositories): In the good case one doesn't understand what %DATA_DIR%
is, and in the bad case one may think it is an environment variable like
on Windows.

Solve this by just telling it is installation-depended and
explanation how to display it in link-parser.