MSVC: Fix get_dictionary_dir() to use LG DLL address
Recently a possibility of relative DICTIONARY_DIR has been added to the
MSVC code. However, it is relative to the location of the link-parser
executable, but the user may use programs in other locations, that use the
LG library.

This change makes DICTIONARY_DIR (if it is a relative path) relative to
the location of the LG DLL.