append_utf8_char(): Avoid gcc8 warning -Wstringop-truncation on strncpy
Avoid this warning on strncpy(buf, mbs, n):
warning: ‘strncpy’ destination unchanged after copying no bytes [-Wstringop-truncation]

There is no bug if n==0, and this function is anyway never be called with
*mbs=='\0'. So assert(*mbs!='\0') at the start could be used to
suppress this warning.

However, instead of adding this slight overhead, and maybe cause a
future run-time surprise if code that uses append_utf8_char() is added
or changed, The fix here is to use memcpy() instead of strncpy() (it
should also be slightly more efficient).