build-disjuncts.c: Arrange for partial pool memory free
While building the clauses, temporary clauses/Tconnectors are used.
The original code freed them.

free them using the POOL_FREE feature, which can free/reuse
individual pool elements.

However, as the whole Clause and Tconnector pools are freed anyway at
the end, this intermediate free just adds a significant overhead, while
not saving (relatively) much memory even for big sentences. So this
code is commented out (but it has been checked with ASAN).

Note that this code needs the definition POOL_FREE in memory-pool.h,
which is not defined for now.