prune.c: Matching shallow connectors: Stop earlier
Connectors cannot match when both are non-shallow.
Now the list of connectors contains mixed shallow/non-shallow, and the
whole list needs scanning, when just before the match try False is
returned as the match result.

Implement the following for stopping earlier in the
left/right_table_search() matching (possible_connection()) loop;

1. On power_table_new(), arrange for the shallow connectors to be first
in the table, by inserting them last (they are inserted in reverse
order).

2. On left/right_table_search(), if we are matching a non-shallow
connector, stop when we don't have any more shallow ones to try.

Note that in a previous commit cleanup_table() got changed not to
change the table order, a thing that is critical for this change.