preparation.c: 2 bug fixes: Add check for alternatives
Fix bug 1:
This code ensures that the same trailing connector sequences which
belong to different alternatives, will get different suffix_id's.
It does that by prepending the gword number(s) of each disjunct to each
connector (disjuncts of different alternatives result from different
gwords).

Fix bug 2:
Indirectly, it also fixes a bug in prepending the word number to the
connector sequence string: It is currently prepended to the whole
string, so only the first (shallow) connector uses it. This may result
in random corrupted linkages if, as a result of this bug, the same
[shallow connector and different deep connectors] pairs have the same
hash value (I didn't observe such a thing, and hence this bug has left
unnoticed in the initial implementation).

The said gword prepending is done for each connector, so it fixes the
said bug and the code to prepend the word number is removed here.