mk_parse_set(): Optimize Rmatch processing
Avoid the Rmatch processing if none of the ls components exists, unless
this is because le==NULL.

This is possible because the loops below that use the rs components
needs at least one of the ls components (first loop), or need le==NULL
(last loop).

This is similar to the optimization in do_count().