SAT parser: Disable the extra pp-pruning code
The purpose of this code is to add restrictions to prevent solutions
that violate a certain pp constraint. This makes finding a good
solution faster.

However, a recent dict change "commit: More locative cases.") reveals
that this check is too restrictive, as this sentence doesn't parse any
more:

--
I think one of the chief differences between the main Filipino language,
which is Tagalog, and English, when I hear Tagalog speakers speaking
English, is that they tend to place the stress on the wrong syllable of
a word or they stress the wrong word in the sentence or in the question.
--

Just disable this check, as the whole SAT parser power-prune() and
pp_prune() are planned to be disabled soon when new code that uses the
classic parser power-prune() and pp_prune() is ready.

Note that this sentence needs !cost=4 in order to get parsed with the
SAT parser, due to some cost cutoff incompatibility it has with the
classic parser. The said new code will make unneeded also the SAT
parser cost cutoff handling, so this problem is left unfixed for now.