Sentence_s: Add connector_suffix_id
In order to be able to implement a calculate the connector suffix_id's
for null_count>0 incrementally (in case a parse with null_count==0 has
already been done for the same sentence), keep its String_id.

This addition also speeds up the calculation in case only a parse with
no nulls is needed (like in batch run or when a sentence is correct).