Use 1-based line numbers
Lines in messages are numbered from 0, but the common convention is that
the first line is number 1. The reported line may depend on whether there
had been a successful link-advance() before the error, and maybe on
other things, and there may be a need to tweak the line number in some
messages.