Unify setting connector suffix_id with connector packing
This is needed for the new power_prune() (WIP), which shares sentence
connectors in a way which is incompatible with the current separate
setting of suffix_id.

Also:
- Refactor into separate functions.
- Shorten some variable names to prevent long lines that need folding.
- Improve comments: Fix comment rot; more exact and detailed explanations.
- Move all the suffix_id setup code out of preparations.c to
disjunct-utils.c, since the suffix_id encoding/packing/sharing are now
done after the pruning, and it is a good idea to perform that for the
SAT parser too (a future work).