Implement memory pools for disjuncts and connectors
Add a first argument to connector_new() and build_disjuncts_for_exp()
for the memory pools. If it is NULL, memory pools are not used.
This is done for use from the SAT-parser, which doesn't use memory
pools.

FIXME: Update the SAT-parser to use memory pools, or alternatively
create disjuncts and connectors directly (which looks simpler than
the current code).