Fix a double pool_delete() of disjuncts/connectors + missing free()
Why this problem happens:
pack_sentence() pack the disjuncts/connectors into its own continuous
memory block and frees their memory-pool.
On one-pass-parse when parse with nulls is needed, restore_disjuncts()
restores the original disjuncts and frees the current ones. But there
are none because they are in a memory block now, so the pool_delete()
fails. Instead, the memory block needs to be freed.

Fix it by
1. Setting sent->Disjunct_pool to NULL to signify that no pool_delete()
of disjuncts/connectors is needed.
2. Instead of freeing memory in restore_disjuncts(), call
free_sentence_disjuncts() since it knows about the memory block
and memory-pool.