tests.py: Add a test to checks memory allocation
It doesn't assert anything useful, but if there is a memory management
problem with the interaction of memory encoding/packing with the
one-step-parse disjunct save/restore, a crash can happen, or a memory
leak (a leak would be detected on script exit).

The test sentence is just one that has minimal parsing overhead.
short_length=1 is used to lower the overhead.