Disable sharing the count table on one pass parses
When doing one pass parses and also suffix_id's are generated (the
sentence is long enough), the count table of the null_count==0 is shared
with higher null counts.
The idea was to use the fact that the even though the power_prune()
results in the case of null_count==0 and null_count>0 are different,
since the same string-id table is used to enumerate them, the same
connector trailing sequence will always get the same number, so
the connector pair count results of the null_count==0 can be reused by
sharing the table.

But something seems to be bad in this idea or its implementation, since
a few sentences (14) in the "fixes" corpus has less linkages when this
idea is used and trailing encoding is forced (they are short enough).
For some reason, all of these linkages happen to be with P.P. violations.

The solution for now: Disable this idea until the reason is found and
fixed.