Implement a new pp_prune()
For the problems of the current one, see issue #912.

This implementation doesn't fix the deletion from the cms table,
since this has no actual effect on more pruning with the current
4.0.knowledge definitions.

The bug that miss later rules is fixed here, but this usually causes
only a few more deletions since the most probable rules that can be
violated are defined at the start.

Do the actual disjunct deletion in power_prune() (which may delete many
more additional disjuncts as a result).

Implement an easy debug facility (PP_PRUNE_DEBUG).