pp_prune(): Implement efficiency hack comment for rule_satisfiable()
The comment (originally from version 3) mentions that once
rule_satisfiable() is false, it continues to be false.

In our case, since we don't delete connectors from the cms table,
if rule_satisfiable() is true it continues to be true.

So once its result is known, remember that and act accordingly.