wordgraph_path_append(): Fix a bug that had no bad effect
When this function is called on the first word slot, it initializes
the wordgraph path. In that case it is called with a NULL gword to add
to the path, which should have been ignored.

The bug is that it actually adds it. But because a NULL Gword serves
as a termination indication of the Gword list, it just serves as a list
terminator (so there are 2 NULL terminators) and no harm happens.

Fixes:

1. Fix that by skipping adding a word if it is NULL.
2. After this fix, another problem happens - the assert() of duplicating
a NULL path is triggered, since "path" remains NULL.
Since in that case the duplication is just to nullify the new path,
fix this bu unifying the checks for path initialization.

This problem was detected while converting the wordgraph path to use
vectors with element count, so a NULL element is just another element
and a first NULL element is destructive.