remove_empty_words(): Optimize call to gwordlist_end()
These call are costly because vc_vector doesn't keep a pointer to the
end and instead it uses a multiplication on each call. The compiler
cannot optimize that in all cases because the result may theoretically
be change (here there is a call to gwordlist_next() that the compiler
cannot assume anything about it).