left_connector_list_update(): Bug fix - ub may be too big
It may be equal to sent_length, and refer beyond the end of
the l_table (will actually refer to r_table[0]).

Because it refers to a valid address, no memory access fault may happen.
It is also undetected by ASAN etc.

In principle it may cause wrong results due to the lookup in the wrong
table.