power_prune(): Don't process again good jets
For long sentences such as "And Yet [...]", it saves ~30% of
the total power_cost (~90% in the WIP tracon-based power_prune()).

Since power_prune() takes now a fraction of a second even for long
sentences like that one, this doesn't save much CPU. But the intention
is to add several more pruning methods so this is an important saving.

NOTE: For the current version (jet-based power_prune()) the position of the
pass_number check is not optimal, as this change was back-ported from
the tracon-based power_prune().