eliminate_duplicate_disjuncts(): Keep the original disjunct order
This is implemented by using an added Disjunct field for the dup table
next element. With this added filed, the Disjunct struct is 64 bytes
instead of 56, which is a good thing (else 2 cache lines are referred
for most Disjunct accesses).

By not overwriting the Disjunct next field, it is possible to discard
disjuncts on the fly, thus there is no need for an additional loop
for constructing a new disjunct list.