tracon_set_reset(): "adaptive" table size
It is unknown in advance how many tracons each word has so the hash
table could be sized accordingly to prevent the overhead of growing it.

This change supposes the number is bigger on the middle words, and tries
to have a better guess than the lower limit (initially set by
tracon_set_create()) by using a similar table size to the previous
words increased or decreased slightly.

A test indicated an increased performance (just using a constant big
table size reduces the performance).

FIXME: Apparently more tuning can be done regarding the initial table size.