power_prune(): Optimize pruning by given null count
Extend the null_count==0 optimization to any null count.
Take the null count from an added argument instead of from the options,
to prevent the need to temporarily mangle min_null_count.

If the null_count argument is MAX_SENTENCE, the optimization is disabled
because the relevant condition is never true.