Disjunct_struct: Rearrange fields, add future fields and comments
- Add fields for incremental connector encoding
They will allow comparing previous and current disjunct lists in order
to preserve the tracon IDs of common disjuncts (needed for reusing the
previous count table when parsing with an increased null_count).

Group fields together according to their use.
The use of 2 unions is just because there is no guarantee that different
pointers have the same size (even so this may be the actual practice for
pointers other than to void or to char).