Reading dictionary: Use memory pools for expressions
Add the following:
Exp and E_list memory pools in struct Dictionary_s.

Remove the following as a result of using memory pools for expressions:

1. The "next" field of Exp_struct.
2. free_Exp_list() / free_Elist().
3. struct Exp_list_s.

Use a pool size of 32 for the affix dict (more than enough
for the biggest dict in the current languages(.

Use a pool size of (1<<13) for the main dict (fastest than e.g. (1>>12),
(1<<14) and even (1<<17) which uses only a single malloc() for the "ru" dict).
(This may change across OSs.)