sat-solver: Fix generate_conjunct_order_constraints () to handle long exp chain
Fix issue #932.

Due to the way the ordering constrains are generated, they are actually
not transitive as clammed in section 2.4 of the SAT-parser paper
(Conjunction order constraints). So invoking
generate_conjunct_order_constraints() for each pair of a long expression
chain (length > 2) misses necessary connector ordering conditions. As a
result, solutions in which connectors are connected in wrong order are
produced.

Repeat by:
---
*A few times may I have taken cocaine, but I inhaled at no time
*I saw make some changes in the program
*Joe doesn't matter what Ted does
*Stravinsky was in Paris that Debussy first heard Balinese music
*The big mind on everybody's question is who killed OJ
---

Work around this bug by converting the expression, on the fly, to
a 2-component chain, just before invoking
generate_conjunct_order_constraints().