Encode incrementally + share previous count table, prune per null_count
Encode incrementally + share previous count table:
The incremental encoding allows keeping the old tracon IDs, which in
turn allows reusing of the previous connector-pair count table (which is
keyed by tracon IDs) in the next parse (with a higher null_count).

Power-prune per null_count:
Before each parse with an increased null_count, perform power_prune()
optimized for the current null_count. This generally increases performance
because the recently rewritten power_prune() is relatively fast.

FIXME (WIP): Building the fast-matcher tables is needed after pruning,
and for "Pathological" sentences with a huge amount of disjuncts this
may take more time than saved by the added pruning per null_count.  A
fast-matcher rewrite will solve this problem and will speed parsing by
reducing the overhead in forming the match lists.

Notes:
1. When power_prune() is performed with an increased null_count,
disjuncts are added (relative to the previous power_prune()). If they
would add "criterion" connectors that are examined by pp_prune(), which
in turn add connector-pairs that have zero entries in the previous
connector-pair count table which is now shared, this may lead to missing
linkages. However, such missing linkages have not been observed in the
English "basic" and "fixes" batches, so it is not clear if this can
really happen. In case this is suspected, it is possible to set the
"test" parse_option to "no-count-sharing" to see if more linkages
appear. See discussion at PR #900 which disabled this feature
(reimplemented here).

2. Using link-parser, sharing the count table for null_count==0 is not
done yet by default, but can be forced using "-test=one-step-parse".
The linkage results should be **totally** identical.