java-ini/Makefile.am: Add src include directory link-grammar
After adding #include "api-types.h" to prune.h (for a WIP), the
compilation of java-client.c failed with this error message:

--
  CC       jni-client.lo
In file included from .../tests/link-grammar/api-structures.h:49,
                 from .../tests/bindings/java-jni/jni-client.c:20:
.../tests/link-grammar/parse/prune.h:16:10: fatal error: api-types.h: No such file or directory
   16 | #include "api-types.h"                  // Tracon_sharing
      |          ^~~~~~~~~~~~~
compilation terminated.
--

I marked below the point of failures with "<--".

The include files in jni-client.c:
     20 #include "link-grammar/api-structures.h" <--
     21 #include "link-grammar/error.h"
     22 #include "jni-client.h"
     23 #include "link-grammar/utilities.h"

This was surprising because the offending include files is indirectly
includes from api-structures.h, which itself has includes api-types.h
just fine - here are its include files (in the WIP branch):

     46 #include "api-types.h"
     47 #include "tracon-set.h"
     48 #include "memory-pool.h"
     49 #include "parse/prune.h" <--
     50 #include "string-set.h"

Include files in parse/prune.h:
     16 #include "api-types.h" <--              // Tracon_sharing
     17 #include "link-includes.h"

The reason of the failure is subtle: "parse/prune.h" above (line 49) is
found under the directory of its parent include (api-structures.h, its
directory is "link-grammar") so it is found (include files are first
searched in the directory of their parent file), a thing that is not
correct for it when it is included from parse/prune.h (its directory is
"parse").

Fix it to prevent such surprises in the future if includes are added to
one of the files which are included by jni-client.c.

Implemented here:
1. Add an implicit -I like the rest of the Makefile.am files.
2. Clean up the include paths.