string-set.c hash_string(): Change multiplication constant 7 -> 139
It turns out 7 is a particularly poor choice, resulted in typical
occurrences of frequent tens of collisions. This is because for short
strings the accumulated hush number is too small, resulted in a high
density in the low range of the table.

Use 139, which was found to give a low number of collisions after
checking many numbers.