find_place(): Fix key upper limit
Theoretically with a big collision number it can overflow.
This has not been observed. For some reason it s very rare that it
goes over the size of the table, and if this happens, it is always
by very few entries,

Be on the safe side and ensure it is always within the table.
As said above, the function will rarely be called.