fast-match.c: Add ML_COMPAT to preserve old disjunct order
The new nearest_word sorting results in reverse order of disjuncts in
the mach lists (relative to the order before this commit).
This makes it harder to compare results to the current version (5.6.2)
because the order of linkages with same metric becomes different.

The old order is worse due to 2 reasons:
1. Added overhead to produce it.
2. The disjuncts in the match lists then appear in descending memory
order, which slows accessing them.

So ML_COMPAT is 0 by default.