count.c: Move the check for left connector farthest word
An optimization for long sentences:
Save the call to form_match_list() and the checks afterwards for
words that are too far for the left connector.

Note that the same thing cannot be done for the right connector due to
the extra needed handling even if it doesn't have matches.