Fix strerror_r() alarming warnings on some systems / compilers
My original implementation of using strerror_r() unnecessarily trying to
somewhat mimic the behavior of the XSI version (it doesn't actually use
the returned value). It also unnecessarily used macros to arrange for
compatibility instead of functions.

Some compilers (rightfully) emit warnings on these definitions.

Fix the problems by arranging the compatibility by a function that
doesn't try to be identical to any version of strerror.  Also provide a
function to print a reasonable message in case there is no strerror_r().
With more efforts it could print the error string but I didn't care
enough.