connector_new(): Bug fix: "shallow" is uninitialized on 32-bit systems
This bug causes random results on 32-bit systems.
It doesn't happen on systems with 64-bit pointers because "shallow"
there overlaps with "originating_gword" with is initialized to NULL.

Fix this by setting the zero_out option to true when defining the pool
for connectors. As a side effect, this has a speedup effect on short
English sentences of about 2-2.5% (only ~0.5% for Russian/long English
sentences).