set_condesc_length_limit(): Fix off-by-1 bug in restart_cn
The bug has a potential to cause missing linkages on sentences longer
than short_length.
This may happen if the qsort() order of both UNLIMITED-CONNECTORS
and the condesc table causes the next condesc in the table to
be one that needs an UNLIMITED-CONNECTORS setting (it will be skipped
due to this off-by-1 bug). So the bug depends on the particular
implementation of qsort() - the produced order of records that are equal.

This is what happened on MacOS "Mojave". The length limit of Xp didn't
get set.

How to see the problem:
link-parser -v=101 </dev/null # condesc table connectors w/their length_limit
(Search for Xp).