classic_parse(): Fix setting pruning for linkage_ok=1
Currently there is a bug when max_null_count > min_null_count:
A prune optimization is done (once) with min_null_count > 0, but this
currently not supported by the pruning code.

As a result, a sentence may be parsed with one more null than needed.

There is no example in the current corpus batches (and I also don't have
an example) for a wrong parsing due to that.

This is a temporary fix to show how it should have been implemented in
the first place, because a better support of pruning for islands_ok=true
with nulls is planned.