assert_same_disjunct(): Enable only on DEBUG_lisjuncts
When morphology=0 and LL links exists, the debug verification assert
fails because the disjunct string doesn't contain then LL links. Since
Parse_options is not available here (the caller doesn't have access to
them) the this debug cannot be skipped on morphology=0, and unexpected
assert failures may happen on debug.

Solve that by enabling the verification only when DEBUG_lisjuncts is
enabled.