tracon_sharing_s: New field is_pruning
Currently the existence of tracon_list is used to differentiate between
a tracon_sharing descriptor for pruning or parsing. This is ambiguous,
as tracon_list is not created for short-enough sentences.
It happened to work by chance but doesn't work fine with future changes.

Define a field "is_pruning" to differentiate between these 2 uses.