Remove count sharing
The idea of count sharing was to speed up parsing with nulls by sharing
the count table between parsings with increasing null counts. For that
to work it is needed to use the same tracon_id's for the same tracons.

This had a nice speedup for parsing with nulls. However, more and more
bugs got found when implementing a more aggressive pruning (but no
examples of problematic existing batch sentences in the latest release).
Fixing them increased the overhead and more bugs have discovered.  It
turned out that this idea is not sound. It is also very hard to
understand and maintain.

Its speedup potential is limited (tens of percents), and it prevents
implementing changes that would increase speedup several times more than
that.

So this idea is dropped until a full-proof way to do that is found.