link-parser.c: Add "@" test to suppress test message
When using -test=something to make a test using a feature (e.g.
-test=one-step-parse), warning are issued that test mode is enabled
(to mark output as using specific tests).

In batch mode only one such warning is issued, but benchmarks that use
-test=... and don't use batch mode will emit a warning per sentence,
a thing that skewed the timing measurement.

Solve it by adding a special test "@" that totall suppress such
warnings.