can_form_link(): Bug fix and comment fix
Fix a wrong returned value that can happen when the connector string is
shorter than the selector string.  This case, listed in the function
comments, is handled wrongly (false is returned):
 *     s="Xa*#"; t="Xa";   e=s[2]; // *e == '*'; // Returns true;

Also fix the comment for another case, in which the indicated result was
incorrect.

It has no impact on the correctness of the parses (detailed check on the
`basic` and `fixes` batches).