read-dict.c: Fix handling of missing ';' at end of dict
It sometimes produces a wrong error message:
UTF8 char is too long
	 Line 17, next tokens:
(Nothing after "next tokens:".

The problem is that when get_character() returns c[0]='\0'
it returns to the loop at link_advance() with ok=true, and then
is_space(c[0]) exist the loop.

The program arrives to:
    if ('\0' == c[1] && char_is_special(c[0]))
which has 2 bugs:
1. c[1] is beyond the final '\0' that is put at the end of dict buffer when it
   is read. It may have some garbage in it and different things may
   happen.
2. char_is_special(c[0]) returns true on '\0' even though this was not
   the intention (since '\0' is found at the end of the string it
   scans).

Since only ';' is missing, this is a position which a cost may apply, so
if (is_number(dict->token))
is called on dict->token which is "\0" at that point.
It has a similar bug as char_is_special() and returns true, but
there is nothing there and the cost is decoded as 0.

The next call to get_character() then tries to read after the buffer
end, gets some garbage and then "UTF8 char is too long".

Fixed by returning true in link_advance() on c[0]=='\0',
and false in is_number of "\0" argument.

It may be that there are more EOF handling problems (there is a
check for the constant EOF which doesn't exists in the dict buffer).