link-parser.c: Bug fix: Don't try 2nd step if 1st one timeouts
If !null=1, since no timeout check is done after the first step, the
second step will be performed. It restarts the timer and if it
doesn't timeout then there will be no indication that a problem happened.

For example, the result can be a parse with null count 1 even though
there is a full linkage. If the sentence parse time is about equal to
the timeout setting, the results may not even be consistent.

Another problem is that the timeout setting is not obeyed.