link-parser.c: Allow setting -verbosity, -debug & -test for dict debugging
This ability was lost in the recent panic-mode changes, due to the need
to read the panic cost setup from the dict and then saving it as a
default along with certain other options. This prevented the processing
of parse options from the command line before opening the dict (s they
would not kept as default ones).

Fix it by processing of the debug options (only) before opening the
dict.