Fix category memory leak when parsing is skipped
If the pruning stage finds that the null count is greater than 0,
then the parse step is skipped and the disjunct memory block is
discarded. However, in generating mode these disjuncts include
categories which have to be freed.

Fix this by freeing them. A new function free_categories() is used.
Call this function from sentence_delete() too instead of the category
freeing loop there.