Disable SAT solver by default.
The SAT solver code appears to be not just a little slower, but a lot
slower in all situations. So, for me, `corpus-fix-long.batch` parses
in a dozen seconds normally, and took over 25 minutes w/ SAT, before
I just killed it. Is that a bug!?

It is 5x slower on corpus-basic.batch
It is 4x slower on corpus-fixes.batch
It is more than 8x slower on Pride & Prejudice (still not done!)