ParseOptions(): Fix memory deallocation
Using __del__ to delete the C parse_options object (_obj) is wrong, as
it (_obj) is deleted if the refcount of the ParseOption object gets to
zero even if the refcount of _obj itself is >0, e.e. it is still needed.
This happens e.g. when clinkgrammar.some_function(..., ParseOptions()._obj)
is invoked (a reference to stale memory would result in this case).