tests.py: YGenerationTestCase: Reset generation mode at test end
The generation mode is not getting reset if the SQL dict cannot be
opened at "test_getting_linkages_sql_dict". This happens because
unittest then aborts the test before evaluating the last argument of
Sentence(), which, if invoked, reset the generation mode by creating en
empty "test" option.

Fix it using tearDownClass(), which is always invoked at test-class end.

This problem happens on the MSVC Windows built because it doesn't include
sqlite support.