dyn_print_disjunct_list(): Don't do repeated alloca()
This may cause stack overflow.
Observed on big disjunct lists when ASAN is enabled (which adds its own
stack overhead per call).

Call alloca() for memory that is big enough for the current dicts,
However, to be on the safe side, call alloca() again if it is turned
out not to be enough.