form_match_list(): Set the cached right count index
A cached right match list may be partially used due to no "left
matching" (see "lc optimization"). So the ordinal number of a right match
list disjunct cannot be used (in do_count()) as an index into the cached
right count array.
To overcome that, keep the needed index in the right list disjuncts.