lrcnt_cache_match_list(): If cannot allocate match list cache, just return
Currently the max. match list cache size is hard coded in
alloc_fast_matcher() to a very big value. In case the match list to be
cached is bigger, just don't cache it.

How to check: Change match_list_pool_size (in alloc_fast_matcher()) to a
small value (say 10) and test with a long sentence.