list_random_links(): Fix a misleading loop exit condition
This loop can never reach the exist condition (pc != NULL) because
(new_index == num_pc) must be true before or at the last element,
because the elements has just been counted.

It is misleading because if this was a valid exit condition, there would
be a null dereference in issue_links_for_choice() which uses pc.

This fix also makes the loop faster.