Fix conditions for "Consider increasing the linkage limit".
The condition (sent->num_valid_linkages == 0) is not needed because
it is already checked to be >0 and it cannot be <0.
The condition (PARSE_NUM_OVERFLOW >= sent->num_linkages_found) is
incorrect.
The message is to be issued if "All examined linkages had P.P. violations"
so add this check.
For now don't issue this message in generation mode (need to check
whether to do that).