Change linkage limit to match link-parser command
This allows the output of the python code to be directly compared to
what link-parser is printing. It also allows the python unit tests
to pass, after the changes to be made in the next commit. Basically,
the spell-check test failed, because the desired spell correction
never showed up in the smaller sample.  Because !spell=99 was set
and !max-linkages=100 and so random sampling kicked in.