Fix russian dict to accomadate new cost computations
The pull req #1456 broke the existing russian dict.
This fixes it.  Seems that a cost cutoff of 3.1 is too low,
as it rejects many good linkages that come from
[[FOO]] & [[BAR]] disjuncts. a cutoff of 4.1 allows these,
while rejecting the cost-6 linkages that ruin corpus-basic.

As of now, I get 45 errors in ru/corpus-basic.batch which is
two less than the 47 that I got, before merging #1450