Provide a superior estimate for the xtable size.
From what I can tell, the tracon pool and the pset pool sizes are
almost (not quite) identical. Because... there should be a parse
set per link type. At least, that would be the case if each link type
is used exactly once in a sentence. (Right? If used more than once,
then there would be multiple parse-sets for that link type...)

(I have not measured this for English; only for MST, which has its
own peculiar way of doing things.)